fix: bulk tags will work now with project permissions
