+++ src/lib/routes/admin-api/tag.ts
@@ -23,7 +23,6 @@ import {
 } from '../../openapi/spec/tag-with-version-schema';
 import { emptyResponse } from '../../openapi/util/standard-responses';
 import FeatureTagService from 'lib/services/feature-tag-service';
-import { TagsBulkAddSchema } from '../../openapi/spec/tags-bulk-add-schema';
 import { IFlagResolver } from '../../types';
 
 const version = 1;
@@ -74,7 +73,7 @@ class TagController extends Controller {
             method: 'post',
             path: '',
             handler: this.createTag,
-            permission: UPDATE_FEATURE,
+            permission: NONE,
             middleware: [
                 openApiService.validPath({
                     tags: ['Tags'],
@@ -88,20 +87,7 @@ class TagController extends Controller {
                 }),
             ],
         });
-        this.route({
-            method: 'put',
-            path: '/features',
-            handler: this.updateFeaturesTags,
-            permission: UPDATE_FEATURE,
-            middleware: [
-                openApiService.validPath({
-                    tags: ['Tags'],
-                    operationId: 'addTagToFeatures',
-                    requestBody: createRequestSchema('tagsBulkAddSchema'),
-                    responses: { 200: emptyResponse },
-                }),
-            ],
-        });
+
         this.route({
             method: 'get',
             path: '/:type',
@@ -206,21 +192,6 @@ class TagController extends Controller {
         const { type, value } = req.params;
         const userName = extractUsername(req);
         await this.tagService.deleteTag({ type, value }, userName);
-        res.status(200).end();
-    }
-
-    async updateFeaturesTags(
-        req: IAuthRequest<void, void, TagsBulkAddSchema>,
-        res: Response<TagSchema>,
-    ): Promise<void> {
-        const { features, tags } = req.body;
-        const userName = extractUsername(req);
-        await this.featureTagService.updateTags(
-            features,
-            tags.addedTags,
-            tags.removedTags,
-            userName,
-        );
         res.status(200).end();
     }
 }
